+++ server/models/meetups.js
@@ -1,6 +1,8 @@
 import fs from 'fs';
 import rsvpData from '../../data/rsvp.json';
 import meetupdata from '../../data/meetups.json';
+import questiondata from '../../data/questions.json';
+
 // GET MEETUP LIST
 let meetupsList = [];
 try {
@@ -14,7 +16,13 @@ try {
 } catch (err) {
     rsvpsList = [];
 }
-
+// GET QUESTION LIST
+let questionsList = [];
+try {
+    questionsList = questiondata;
+} catch (err) {
+    questionsList = [];
+}
 // Create mettups
 function addMeetup(data) {
     fs.writeFile('././data/meetups.json', JSON.stringify(data, null, 2), (err) => {
@@ -27,7 +35,12 @@ function addRsvp(data) {
         if (err) throw err;
     });
 }
-//
+// Create Question
+function addQuestion(data) {
+    fs.writeFile('././data/questions.json', JSON.stringify(data, null, 2), (err) => {
+        if (err) throw err;
+    });
+}
 export {
-addMeetup, meetupsList as meetups, addRsvp, rsvpsList as rsvps
+addMeetup, meetupsList as meetups, addRsvp, rsvpsList as rsvps, addQuestion, questionsList
 };
