+++ server/routes/meetups.js
@@ -6,12 +6,14 @@ const router = express.Router();
 // Get Meetups
 router.get('/', controller.get_meetups);
 // Get Upcoming Meetup
-router.get('/', controller.get_upcoming_meetups);
+router.get('/upcoming', controller.get_upcoming_meetups);
 // Get Upcoming Meetup
 router.get('/:id', controller.get_specific_meetups);
 // Post Meetup
 router.post('/', controller.post_meetups);
 // Post Rsvp
 router.post('/:id/rsvps', controller.post_rsvp);
+// Post question on specific meetup
+router.post('/:id/questions', controller.post_question);
 
 export default router;
